Merging

+++ src/context/like-context.js
@@ -1,46 +1,52 @@
 import { useContext, useState, createContext } from "react";
 import { useNavigate } from "react-router-dom";
+import { useAuth } from "./auth-context";
 import axios from "axios";
 
 const LikedContext = createContext();
 
 const LikedProvider = ({ children }) => {
+  const { loggedIn } = useAuth();
   const navigate = useNavigate();
   const [likedList, setLikedList] = useState([]);
 
   const addToLikeList = async (video) => {
-    try {
-      const token = localStorage.getItem("userToken");
-      const response = await axios.post(
-        "/api/user/likes",
-        { video },
-        { headers: { authorization: token } }
-      );
-      if (response.status === 201) {
-        setLikedList(response.data.likes);
+    if (loggedIn) {
+      try {
+        const token = localStorage.getItem("userToken");
+        const response = await axios.post(
+          "/api/user/likes",
+          { video },
+          { headers: { authorization: token } }
+        );
+        if (response.status === 201) {
+          setLikedList(response.data.likes);
+        }
+      } catch (error) {
+        console.log("error", error);
+        navigate("/login");
       }
-    } catch (error) {
-      console.log("error", error);
-      navigate("/login");
     }
   };
 
   const removeFromLikedList = async (videoId) => {
-    try {
-      const token = localStorage.getItem("userToken");
-      const { data, status } = await axios.delete(
-        `/api/user/likes/${videoId}`,
-        {
-          headers: {
-            authorization: token,
-          },
+    if (loggedIn) {
+      try {
+        const token = localStorage.getItem("userToken");
+        const { data, status } = await axios.delete(
+          `/api/user/likes/${videoId}`,
+          {
+            headers: {
+              authorization: token,
+            },
+          }
+        );
+        if (status === 200) {
+          setLikedList(data.likes);
         }
-      );
-      if (status === 200) {
-        setLikedList(data.likes);
+      } catch (error) {
+        console.log("error", error);
       }
-    } catch (error) {
-      console.log("error", error);
     }
   };
   return (
